refactor: clean up folder structure
