+++ packages/frontend/src/components/VisualizationConfigs/ChartConfigPanel/ConfigTabs/index.tsx
@@ -1,12 +1,12 @@
 import { MantineProvider, Tabs } from '@mantine/core';
 import { memo, useMemo, type FC } from 'react';
 import { useVisualizationContext } from '../../../LightdashVisualization/VisualizationProvider';
+import { themeOverride } from '../../theme';
 import { Axes } from '../Axes';
 import { Grid } from '../Grid';
 import { Layout } from '../Layout';
 import { Legend } from '../Legend';
 import { Series } from '../Series';
-import { themeOverride } from './mantineTheme';
 
 export const ConfigTabs: FC = memo(() => {
     const { itemsMap } = useVisualizationContext();
