+++ packages/frontend/src/components/VisualizationConfigs/ChartConfigPanel/Layout/index.tsx
@@ -24,7 +24,7 @@ import FieldSelect from '../../../common/FieldSelect';
 import MantineIcon from '../../../common/MantineIcon';
 import { isCartesianVisualizationConfig } from '../../../LightdashVisualization/VisualizationConfigCartesian';
 import { useVisualizationContext } from '../../../LightdashVisualization/VisualizationProvider';
-import { MAX_PIVOTS } from '../../TableConfigPanel/GeneralSettings';
+import { MAX_PIVOTS } from '../../TableConfigPanel/General';
 import { AddButton } from '../common/AddButton';
 import { Config } from '../common/Config';
