+++ packages/frontend/src/components/VisualizationConfigs/TableConfigPanel/General/ColumnConfiguration.tsx
@@ -7,9 +7,9 @@ import {
     IconLockOpen,
 } from '@tabler/icons-react';
 import React, { useState } from 'react';
-import MantineIcon from '../../common/MantineIcon';
-import { isTableVisualizationConfig } from '../../LightdashVisualization/VisualizationConfigTable';
-import { useVisualizationContext } from '../../LightdashVisualization/VisualizationProvider';
+import MantineIcon from '../../../common/MantineIcon';
+import { isTableVisualizationConfig } from '../../../LightdashVisualization/VisualizationConfigTable';
+import { useVisualizationContext } from '../../../LightdashVisualization/VisualizationProvider';
 
 const ColumnConfiguration: React.FC<{ fieldId: string }> = ({ fieldId }) => {
     const { pivotDimensions, visualizationConfig } = useVisualizationContext();
